[MASTERDIET-12] Post review fixes.
