+++ src/main/java/pl/agh/edu/master_diet/config/SecurityConfig.java
@@ -92,7 +92,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .authenticationEntryPoint(new RestAuthenticationEntryPoint())
                 .and()
                 .authorizeRequests()
-                .antMatchers("/ping", "/actuator/*", "/plan").permitAll()
+                .antMatchers("/ping", "/actuator/*").permitAll()
                 .antMatchers("/",
                         "/error",
                         "/favicon.ico",
