+++ src/main/java/pl/agh/edu/master_diet/service/converter/ConversionServiceImpl.java
@@ -3,7 +3,9 @@ package pl.agh.edu.master_diet.service.converter;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.agh.edu.master_diet.core.model.database.UserPlan;
 import pl.agh.edu.master_diet.core.model.rest.userPlan.UserParametersRequest;
+import pl.agh.edu.master_diet.core.model.rest.userPlan.UserPlanResponse;
 import pl.agh.edu.master_diet.core.model.shared.UserParameters;
 
 @Service
@@ -11,9 +13,15 @@ import pl.agh.edu.master_diet.core.model.shared.UserParameters;
 public final class ConversionServiceImpl implements ConversionService {
 
     private final UserParametersConverter userParametersConverter;
+    private final UserPlanConverter userPlanConverter;
 
     @Override
     public UserParameters convert(UserParametersRequest request) {
         return userParametersConverter.createFrom(request);
+    }
+
+    @Override
+    public UserPlanResponse convert(UserPlan userPlan) {
+        return userPlanConverter.createFrom(userPlan);
     }
 }
