+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:postgresql://localhost:5433/master-diet-db
+    url: jdbc:postgresql://master-diet-db:5432/master-diet-db
     username: postgres
     password: postgres
     driver-class-name: org.postgresql.Driver
